Fix more SDL2 issues
  - ContinueMenuState:
    * Update onMouseWheel method to match nomlib's SDL2 API
    * Match nomlib's SDL2 API for onUserEvent argument types

  * GameOverState: Match nomlib's SDL2 API for onUserEvent argument
types

+++ src/states/GameOverState.hpp
@@ -63,7 +63,7 @@ class GameOverState: public nom::IState
     void onKeyDown ( nom::int32 key, nom::int32 mod, nom::uint32 window_id );
     void onMouseLeftButtonDown ( nom::int32 x, nom::int32 y, nom::uint32 window_id );
     void onMouseWheel ( nom::int32 x, nom::int32 y, nom::uint32 window_id );
-    void onUserEvent ( nom::uint8 type, nom::int32 code, void* data1, void* data2 );
+    void onUserEvent ( nom::uint32 type, nom::int32 code, void* data1, void* data2 );
 
     void update ( float delta_time );
     void draw ( nom::IDrawable::RenderTarget target );
